+++ src/app/game/game.component.html
@@ -10,9 +10,9 @@
       <td><span class="btn" (click)="customGame()">Custom</span></td>
     </tr>
     <tr *ngIf="displayInputs">
-      <td><b>Rows: </b><input type="number" min="1" [(ngModel)]="rowsInput" (input)="updateBoard()"></td>
-      <td><b>Columns: </b><input type="number" min="1" [(ngModel)]="columnsInput" (input)="updateBoard()"></td>
-      <td><b>Mines: </b><input type="number" min="1" [(ngModel)]="minesInput" [max]="maxMines" (input)="updateBoard()"></td>
+      <td><b>Rows: </b><input type="number" min="1" [(ngModel)]="input.row" (input)="updateBoard()"></td>
+      <td><b>Columns: </b><input type="number" min="1" [(ngModel)]="input.column" (input)="updateBoard()"></td>
+      <td><b>Mines: </b><input type="number" min="1" [(ngModel)]="input.mine" [max]="maxMines" (input)="updateBoard()"></td>
     </tr>
   </table>
   <classical-board
